separate auth schemas
